move slack related bot code to a sub module
