+++ bot/slack/slack_test.go
@@ -1,4 +1,4 @@
-package bot
+package slack
 
 import (
 	"context"
